requirements.txt
pandas==2.2.0
streamlit==1.33.0
Pillow==10.2.0
xgboost==2.0.3
scikit-learn==1.4.2
path==16.14.0
